python/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
black==23.12.1
certifi==2023.11.17
click==8.1.7
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
isort==5.13.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
replicate==0.22.0
sniffio==1.3.0
typing_extensions==4.9.0
